Fix nullPointerExeption and Consistency Predicate problems
When creating/using an object in which the FenixFrameWork controls it is necessary
to use the notation @Atomic (this notation indicates that the method will result in
a database transaction) to avoid exceptions of the type nullPointerExeption.
Relations with multiplicities with the lower bound greater than 0, or with an upper
bound greater than 1 and lesser than * will generate consistency predicates. (Read
https://fenix-ashes.ist.utl.pt/fenixWiki/ConsistencyPredicates), ie all participants in
the relation must have to be created at the same time to prevent consistency
problems in the database.
It has created a static method called setup in the main class to demonstrate the
use of @Atomic and the problem of consistency in the database. The Constructors
of File, User and SuperUser had to be commented